temperature/requirements.txt
streamlit
easyocr
pillow-heif
pandas
Pillow
numpy
gspread
google-auth
